Add new passes, cycle GC, and harden compiler pipeline
Extends the pass pipeline with HmTypeInfer, Inline, LoopUnroll, Exhaustive,
and GcAnnotate passes, fixing dominance correctness in GC annotation so
Release instructions only reference values whose defining block dominates
the return site.

Implements Bacon-Rajan cycle collection in the C runtime with thread-safe
retain/release via a global mutex, preventing memory leaks from cyclic
heap structures.

Fixes operator precedence: comparison operators now bind tighter than
logical AND/OR but looser than arithmetic, matching expected semantics.

Adds module-qualified function calls (e.g. `module.func()`), interpreter
error location tracking for better diagnostics, and fallback to the
interpreter when clang is unavailable during `iris run`.

Replaces direct `parse_module()` calls with `parse_recovering()` across
all compile paths to surface multiple parse errors before aborting.

Hardens the lowerer with undefined-callee detection, for-loop type
validation, and generic type substitution based on parameter types
rather than positional type-param matching.

Adds comprehensive test suites for the debugger, LSP, REPL, and GC
annotation pass. Fixes fuzz targets to compile on stable Rust by
gating `no_main`/libfuzzer behind `cfg(fuzzing)` with a `fn main`
fallback that smoke-tests the corpus.

Annotates stdlib `list()` calls with explicit type parameters to
avoid Infer propagation, and bumps the VS Code extension to 0.3.0.

+++ tests/parse_lower.rs
@@ -32,7 +32,7 @@ def matmul(A: tensor<f32, [M, K]>, B: tensor<f32, [K, N]>) -> tensor<f32, [M, N]
 "#;
 
 const CONST_SRC: &str = r#"
-def always_one() -> f32 {
+def always_one() -> f64 {
     1.0
 }
 "#;
